rename: amazons3, scheduler 패키지 이름 클래스 이름 변경
